move redux logics to container, remove unnecessary files/components
