--- src/components/hooksFormWrapper/index.js
@@ -1,2 +1,0 @@
-export { default as TextFieldWrapper } from './TextFieldWrapper';
-export { default as ToggleBlockWrapper } from './ToggleBlock.js';
